correcciones finales

+++ parcial.py
@@ -48,11 +48,27 @@ def guardar_planilla(archivo, campos):
         salir = input("Desea salir del ingreso de empleados? si/no: ")
     
     try:
-        with open(archivo, "w", newline="") as file:
-            planilla = csv.writer(file)
-            planilla.writerows(lista_empleados)
-            print("\nEl archivo se guardo correctamente.")
-            return
+        archivo_existe = os.path.isfile(archivo)
+
+        if archivo_existe:
+            sobreescribir = input(f"\nEl archivo existe. Desea sobreescribir? si/no: ")
+            
+            if sobreescribir == "si":
+                with open(archivo, "w", newline="") as file:
+                    planilla = csv.writer(file)
+                    planilla.writerows(lista_empleados)
+                    print("\nEl archivo se guardo correctamente.")
+            else:
+                with open(archivo, "a", newline="") as file:
+                    planilla = csv.writer(file)
+                    planilla.writerows(lista_empleados)
+                    print("\nEl archivo se modifico correctamente.")
+        else:
+            with open(archivo, "w", newline="") as file:
+                planilla = csv.writer(file)
+                planilla.writerows(lista_empleados)
+                print("\nEl archivo se guardo correctamente.")
+        return
     except IOError:
         print("\nOcurrio un error con el archivo.")
